+++ changelog.md
@@ -1,5 +1,6 @@
 # v0.1.2
 1. async file & response
+2. static files
 
 # v0.1.1
 1. descorators can aysnc func